Total Row Stylings
The tr.total added a weird blue color. I cleared that out and added a style for the tfoot to (hopefully) make that row stand out. I also noticed this would make 3 "Total" labels, so I added "Gross" and "Net" to the ones in the thead to clarify what they are signifying.