open PSHR paystub page in a new tab
This is important because it uses javascript to create new popups, which get blocked by default in chrome. The block notification is invisible if the page is not a top-level page; we can't let it be an iframe.
Also, this is consistent with the 'ePayroll history' link at the bottom of the page.