add "?NAVSTACK=Clear" to the pshr link
This will hopefully stop the difficult-to-reproduce problem of the user being redirected to the pshr main page, instead of the pay stubs page.
The problem does not show with a clean set of cookies. I think it shows when the user has had some kind of recent interaction with pshr.

This parameter is documented to be "used for context sensitive return to links."
Also, "NAVSTACK=Clear parameter has the undocumented feature of not updating breadcrumbs."

It's not clear to me that this will really fix the problem, but it seemed to fix it in at least one case (for Nick K).