don't use Function.callee
This updates Paul Irish's log wrapper to his most recently published version,
which is simpler than what was here before and doesn't use 'caller' or 'callee'.
These are deprecated and Safari now fails when these are called on a strict mode function,
such as those in easy-xdm.js.

I'm also using the non-minified version, because it's easier to understand.